Rename the files to appropriate names
